# [DNM] Test memory dirtying, COW chunk size